+++ app/widgets/Notification/notification.js
@@ -12,6 +12,33 @@ var Notification = {
     notifs_key : '',
     favicon : null,
 
+    audioCall : null,
+
+    incomingMessage : function ()
+    {
+        if (NOTIFICATION_CHAT) {
+            // From https://free-mobi.org/ringtones/sms/sms-sound-2
+            var tone = new Audio('theme/audio/message.ogg');
+            tone.play();
+        }
+    },
+    incomingCall : function ()
+    {
+        if (NOTIFICATION_CALL) {
+            // From https://www.zedge.net/ringtone/466d15be-8fa0-32a1-b3dc-62c12a86b6da
+            Notification.audioCall = new Audio('theme/audio/call.ogg');
+            Notification.audioCall.addEventListener('ended', function() {
+                this.currentTime = 0;
+                this.play();
+            }, false);
+            Notification.audioCall.play();
+        }
+    },
+    incomingAnswer : function ()
+    {
+        Notification.audioCall.pause();
+        Notification.audioCall.currentTime = 0;
+    },
     inhibit : function(sec) {
         Notification.inhibed = true;