+++ database/migrations/20200327114708_add_notifications_to_users_table.php
@@ -1,0 +1,23 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddNotificationsToUsersTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('users', function (Blueprint $table) {
+            $table->boolean('notificationchat')->default(true);
+            $table->boolean('notificationcall')->default(true);
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('users', function (Blueprint $table) {
+            $table->dropColumn('notificationchat');
+            $table->dropColumn('notificationcall');
+        });
+    }
+}